+++ ec/types/multi.py
@@ -41,3 +41,29 @@ class one_of(CustomType):
     
   def __str__(self):
     return '%s' % '/'.join(self.choices)
+    
+class menu(CustomType):
+  def __init__(self, choices):
+    self.choices = choices
+    
+  def __call__(self, val):
+    try:
+      val = int(val)
+      
+      if not 0 < val < len(self.choices):
+        raise Exception('')
+      
+      return self.choices[val - 1]
+      
+    except Exception as e:
+      raise ValueError('Invalid value.')
+    
+  def __str__(self):
+    ret = 'Select:\n'
+    n = 0
+    choices = self.choices
+    
+    for n in range(0, len(choices) - 1):
+      ret += '\t%s: %s\n' % (n + 1, self.choices[n])
+    
+    return ret